adding runcommand refactor
