+++ api/cms/download_submission.cms.go
@@ -8,7 +8,7 @@ import (
 
 	//jwt "github.com/appleboy/gin-jwt"
 	"github.com/gin-gonic/gin"
-	"github.com/mongodb/mongo-go-driver/bson/objectid"
+	"github.com/mongodb/mongo-go-driver/bson/primitive"
 
 	"github.com/stevens-tyr/tyr-gin"
 )
@@ -43,7 +43,7 @@ func DownloadSubmission(c *gin.Context) {
 
 	//claims := jwt.ExtractClaims(c)
 	//submittedFileName := fmt.Sprintf("%s%s%s%s.tar.gz", c.Param("cid"), c.Param("aid"), claims["uid"].(string), c.Param("sid"))
-	fileID, err := objectid.FromHex(c.Param("sid"))
+	fileID, err := primitive.ObjectIDFromHex(c.Param("sid"))
 	if !tyrgin.ErrorHandler(err, c, 500, gin.H{
 		"staus_code": 500,
 		"message":    "Inavlid submission id.",
