+++ go.mod
@@ -7,7 +7,7 @@ require (
 	github.com/goware/emailx v0.0.0-20171023230436-0bae9679d4e3
 	github.com/h2non/filetype v1.0.6
 	github.com/mongodb/mongo-go-driver v0.2.0
-	github.com/stevens-tyr/tyr-gin v0.0.0-20190131221820-2e35a7dea952
+	github.com/stevens-tyr/tyr-gin v0.0.0-20190201015310-b826c14790ce
 	github.com/stretchr/testify v1.3.0
 	golang.org/x/crypto v0.0.0-20190131182504-b8fe1690c613
 )
